Sep 15 [FIX] : Bug Fixed 'pos_button_visibility'
