+++ pos_button_visibility/README.rst
@@ -1,10 +1,11 @@
-.. image:: https://img.shields.io/badge/licence-LGPL--3-blue.svg
-    :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
+.. image:: https://img.shields.io/badge/licence-LGPL--3-green.svg
+    :target: https://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 
-Pos Button Visibility
+User Wise Button Restrict In POS
 =================================
-Pos button visibility is used to visible the buttons in the pos basis of users and sessions
+User Wise Button Restrict In POS is used to restrict the buttons in the pos
+basis of users and sessions
 
 Configuration
 =============
@@ -14,10 +15,14 @@ Company
 -------
 * `Cybrosys Techno Solutions <https://cybrosys.com/>`__
 
+License
+-------
+General Public License, Version 3 (LGPL v3).
+(https://www.gnu.org/licenses/lgpl-3.0-standalone.html)
+
 Credits
 -------
-* Developers: 	Cybrosys Techno Solutions odoo@cybrosys.com
-                Version 16: Gayathri V @cybrosys
+* Developer: (V16) Gayathri V, Contact: odoo@cybrosys.com
 
 Contacts
 --------
