+++ pos_button_visibility/__init__.py
@@ -1,4 +1,3 @@
-"""pos button visibility"""
 # -*- coding: utf-8 -*-
 #############################################################################
 #
@@ -20,4 +19,4 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-from .import models
+from . import models
