+++ pos_button_visibility/__manifest__.py
@@ -20,33 +20,34 @@
 #
 #############################################################################
 {
-    'name': 'Pos Button Visibility ',
-    'depends': ['base', 'point_of_sale', 'pos_sale', 'pos_loyalty'],
-    'version': '16.0.1.0.0',
+    'name': 'User Wise Button Restrict In POS ',
+    'version': '16.0.1.1.0',
     'category': 'Point Of Sale',
-    'summary': """Pos button visibility is used to visible the buttons in the pos
-     basis of users and sessions""",
-    'description': """pos buttons can be restricted to the users in a specified sessions""",
+    'summary': """User Wise Button Restrict In POS is used to restrict the 
+    buttons in the pos basis of users and sessions""",
+    'description': """pos buttons can be restricted to the users in a 
+    specified sessions""",
     'author': 'Cybrosys Techno Solutions',
     'company': 'Cybrosys Techno Solutions',
     'maintainer': 'Cybrosys Techno Solutions',
     'website': 'https://www.cybrosys.com',
-    'images': ['static/description/banner.png'],
-    'license': 'LGPL-3',
-    'installable': True,
-    'auto_install': False,
-    'application': False,
+    'depends': ['base', 'point_of_sale', 'pos_sale', 'pos_loyalty'],
     'data': {
         'security/ir.model.access.csv',
-        'data/pos_button_visibility_data.xml',
+        'data/pos_buttons_data.xml',
         'views/res_users_views.xml',
     },
     'assets': {
         'point_of_sale.assets': [
-            'pos_button_visibility/static/src/js/refundbutton.js',
-            'pos_button_visibility/static/src/xml/refundbutton_templates.xml',
-            'pos_button_visibility/static/src/xml/rewardbutton_templates.xml',
-            'pos_button_visibility/static/src/xml/numpadwidget_templates.xml',
+            'pos_button_visibility/static/src/js/RefundButton.js',
+            'pos_button_visibility/static/src/xml/RefundButton.xml',
+            'pos_button_visibility/static/src/xml/RewardButton.xml',
+            'pos_button_visibility/static/src/xml/NumpadWidget.xml',
         ]
     },
+    'images': ['static/description/banner.jpg'],
+    'license': 'LGPL-3',
+    'installable': True,
+    'auto_install': False,
+    'application': False,
 }
