+++ pos_button_visibility/data/pos_buttons_data.xml
@@ -1,14 +1,18 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
+<!--    Record for Refund in model pos.button-->
     <record id="refund_pos_buttons" model="pos.buttons">
         <field name="name">Refund</field>
     </record>
+    <!--    Record for Price in model pos.button-->
     <record id="pos_price_buttons" model="pos.buttons">
         <field name="name">Price</field>
     </record>
+    <!--    Record for Discount in model pos.button-->
     <record id="pos_discount_buttons" model="pos.buttons">
         <field name="name">Discount</field>
     </record>
+    <!--    Record for Reward in model pos.button-->
     <record id="pos_reward_buttons" model="pos.buttons">
         <field name="name">Reward</field>
     </record>
