+++ pos_button_visibility/doc/RELEASE_NOTES.md
@@ -3,4 +3,8 @@
 #### 10.07.2023
 #### Version 16.0.1.0.0
 ##### ADD
-- Initial commit for Pos Button Visibility 
+- Initial commit for User Wise Button Restrict In POS 
+
+#### 15.09.2023
+#### Version 16.0.1.1.0
+#### DEBUG
