+++ pos_button_visibility/security/ir.model.access.csv
@@ -1,2 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_pos_buttons,pos.buttons,model_pos_buttons,base.group_user,1,1,1,1
+access_pos_buttons_user,pos.buttons.user,model_pos_buttons,base.group_user,1,1,1,1
