+++ pos_button_visibility/static/description/index.html
@@ -23,9 +23,9 @@
 
     <!-- APP HERO -->
     <h1 style="color: #FFFFFF; font-weight: bolder; font-size: 50px; text-align: center; margin-top: 50px;">
-        Pos Button Visibility</h1>
+        User Wise Button Restrict In POS</h1>
     <p style="color:#FFFFFF; padding: 8px 15px; text-align: center; font-size: 24px;">
-        A module for restricting users to access the pos buttons.</p>
+         This module for restricting users to access the pos buttons.</p>
     <!-- END OF APP HERO -->
     <img src="assets/screenshots/hero.gif"
          style="width: 75%; height: auto; position: absolute; margin-left: auto; margin-right: auto; top: 45%; left: 12%; right: auto;"/>
@@ -108,7 +108,7 @@
 <div class="row"
      style="font-family: 'Montserrat', sans-serif; font-weight: 400; font-size: 14px; line-height: 200%;">
     <div class="col-sm-12 py-4">
-        Pos button visibility is used to disable buttons for users based on the
+        User Wise Button Restrict In POS is used to disable buttons for users based on the
         session.
     </div>
 </div>
@@ -211,34 +211,34 @@
                     <div class="carousel-item" style="min-height: 198.656px;">
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/16.0/myfatoorah_payment_gateway/"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/custom_receipts_for_pos/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/fatoorah.png">
+                                         src="assets/modules/custom_receipts_for_pos.png">
                                 </div>
                             </a>
                         </div>
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/16.0/website_cart_clear/"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/pos_book_order/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/website_cart.png">
+                                         src="assets/modules/pos_book_order.jpg">
                                 </div>
                             </a>
                         </div>
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/16.0/hr_zk_attendance/"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/point_of_sale_logo/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/integration_biometric.png">
+                                         src="assets/modules/point_of_sale_logo.png">
                                 </div>
                             </a>
                         </div>
@@ -247,34 +247,34 @@
                          style="min-height: 198.656px;">
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/16.0/pos_credit_limit/"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/pos_receipt_extend/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/product_brand.png">
+                                         src="assets/modules/pos_receipt_extend.png">
                                 </div>
                             </a>
                         </div>
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/browse?search=pos_custom_percentage_tip_fixed"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/login_pos_direct/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/barcode.png">
+                                         src="assets/modules/login_pos_direct.png">
                                 </div>
                             </a>
                         </div>
                         <div class="col-xs-12 col-sm-4 col-md-4 mb16 mt16"
                              style="float:left">
-                            <a href="https://apps.odoo.com/apps/modules/16.0/product_brand_ecommerce/"
+                            <a href="https://apps.odoo.com/apps/modules/16.0/pos_access_right_hr/"
                                target="_blank">
                                 <div style="border-radius:10px">
                                     <img class="img img-responsive center-block"
                                          style="border-radius: 0px;"
-                                         src="assets/modules/product.png">
+                                         src="assets/modules/pos_access_right_hr.png">
                                 </div>
                             </a>
                         </div>
@@ -427,7 +427,7 @@
 
     </div>
 
-    <!-- END OF END OF OUR SERVICES -->
+    <!-- END OF OUR SERVICES -->
 
     <!-- OUR INDUSTRIES -->
 
@@ -565,7 +565,7 @@
         </div>
     </div>
 
-    <!-- END OF END OF OUR INDUSTRIES -->
+    <!-- END OF OUR INDUSTRIES -->
 
     <!-- SUPPORT -->
     <div class="d-flex align-items-center"
