+++ pos_button_visibility/static/src/xml/NumpadWidget.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates id="template" xml:space="preserve">
-    <!--This is used to disable the restricted buttons -->
+    <!--This is used to disable the restricted Price button -->
     <t t-name="Price" t-inherit="point_of_sale.NumpadWidget"
        t-inherit-mode="extension" owl="1">
         <xpath expr="//div[hasclass('numpad')]//button[hasclass('mode-button')][3]"
