+++ pos_button_visibility/static/src/xml/RefundButton.xml
