+++ pos_button_visibility/static/src/xml/RewardButton.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates id="template" xml:space="preserve">
-   <!-- This is used to restrict the REFUND  button if it is restricted -->
+   <!-- This is used to restrict the reward  button if it is restricted -->
     <t t-name="DisableReward" t-inherit="pos_loyalty.RewardButton"
        t-inherit-mode="extension" owl="1">
         <xpath expr="//span" position="replace">
