+++ pos_button_visibility/views/res_users_views.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
     <!-- Adding pos sessions and buttons to the user form-->
-    <record id="res_users_view_form" model="ir.ui.view">
-        <field name="name">res.users.view.form.inherit.pos.button.visibility
-        </field>
+    <record id="view_users_form" model="ir.ui.view">
+        <field name="name">res.users.view.form.inherit.pos.button.visibility</field>
         <field name="model">res.users</field>
         <field name="inherit_id" ref="base.view_users_form"/>
         <field name="arch" type="xml">
